컴포넌트 넣는 것은 끝났는데 지금 빛 갱신이 안되는거 같음

+++ KMGLib/ShaderShared.hlsl
@@ -21,7 +21,7 @@ cbuffer cbChangeOnResize : register(b2)
 
 struct Light
 {
-    float type; // 0 = directional, 1 = point, 2 = spot
+    int type; // 0 = directional, 1 = point, 2 = spot
 
     float3 position;
     float3 direction;
